stm32: spi: make cs optional
Some chips (like the ad 400x series) dont have a
CS, and use the CNV pin (or the SDI) pins as CS.
Sometimes the CNV need to be held HIGH and not low
while data is read.

For these scenarios make the SPI CS optional so
that the driver will handle it before the spi
operation.

+++ include/no_os_spi.h
@@ -134,7 +134,7 @@ struct no_os_spi_init_param {
 	/** maximum transfer speed */
 	uint32_t	max_speed_hz;
 	/** SPI chip select */
-	uint8_t		chip_select;
+	int32_t		chip_select;
 	/** SPI mode */
 	enum no_os_spi_mode	mode;
 	/** SPI bit order */
